requirements.txt
langchain==0.1.0
langchain-google-genai==0.0.6
langchain-community==0.0.20
chromadb==0.4.22
pandas==2.1.4
PyPDF2==3.0.1
python-dotenv==1.0.0
google-generativeai==0.3.2 